backend/requirements-dev.txt
# backend/requirements.txt
# Framework
fastapi[standard]==0.135.1
uvicorn[standard]==0.41.0

# Base de datos
sqlmodel==0.0.22
asyncpg==0.29.0  # Para PostgreSQL
psycopg2-binary==2.9.9

# Validación y configuración
pydantic==2.12.5
pydantic-settings==2.13.1
python-dotenv==1.0.0
email-validator==2.3.0

# Utilidades
python-multipart==0.0.22
httpx==0.28.1
click==8.3.1
typer==0.24.1

# Logging y monitoreo
sentry-sdk==2.54.0
rich==14.3.3

# Seguridad (opcional, si las necesitas)
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1

#encryptacion
cryptography==46.0.5